+++ packages/store-sync/src/blockLogsToStorage.test.ts
@@ -2,6 +2,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { BlockLogsToStorageOptions, blockLogsToStorage } from "./blockLogsToStorage";
 import storeConfig from "@latticexyz/store/mud.config";
 import { isDefined } from "@latticexyz/common/utils";
+import { TableId } from "@latticexyz/common";
 
 const mockedCallbacks = {
   registerTables: vi.fn<
@@ -34,6 +35,7 @@ describe("blockLogsToStorage", () => {
           if (table.namespace === "" && table.name === "Inventory") {
             return {
               ...table,
+              tableId: TableId.toHex("", "Inventory"),
               keySchema: {
                 owner: "address",
                 item: "uint32",
