+++ packages/store-sync/src/common.ts
@@ -1,13 +1,9 @@
-import {
-  DynamicAbiType,
-  DynamicPrimitiveType,
-  SchemaAbiType,
-  SchemaAbiTypeToPrimitiveType,
-  StaticAbiType,
-  StaticPrimitiveType,
-  schemaAbiTypeToDefaultValue,
-} from "@latticexyz/schema-type";
-import { Address, Hex, getAddress } from "viem";
+import { SchemaAbiType, SchemaAbiTypeToPrimitiveType, StaticAbiType } from "@latticexyz/schema-type";
+import { Address, Hex } from "viem";
+// TODO: move these type helpers into store?
+import { Key, Value } from "@latticexyz/store-cache";
+import { GetLogsResult, GroupLogsByBlockNumberResult } from "@latticexyz/block-logs-stream";
+import { StoreEventsAbi, StoreConfig } from "@latticexyz/store";
 
 export type ChainId = number;
 export type WorldId = `${ChainId}:${Address}`;
@@ -15,9 +11,16 @@ export type WorldId = `${ChainId}:${Address}`;
 export type TableNamespace = string;
 export type TableName = string;
 
-export type TableRecord = {
-  key: Record<string, StaticPrimitiveType>;
-  value: Record<string, StaticPrimitiveType | DynamicPrimitiveType>;
+export type KeySchema = Record<string, StaticAbiType>;
+export type ValueSchema = Record<string, SchemaAbiType>;
+
+export type SchemaToPrimitives<TSchema extends ValueSchema> = {
+  [key in keyof TSchema]: SchemaAbiTypeToPrimitiveType<TSchema[key]>;
+};
+
+export type TableRecord<TKeySchema extends KeySchema = KeySchema, TValueSchema extends ValueSchema = ValueSchema> = {
+  key: SchemaToPrimitives<TKeySchema>;
+  value: SchemaToPrimitives<TValueSchema>;
 };
 
 export type Table = {
@@ -25,23 +28,53 @@ export type Table = {
   tableId: Hex;
   namespace: TableNamespace;
   name: TableName;
-  keySchema: Record<string, StaticAbiType>;
-  valueSchema: Record<string, StaticAbiType | DynamicAbiType>;
-  lastUpdatedBlockNumber: bigint | null;
+  keySchema: KeySchema;
+  valueSchema: ValueSchema;
 };
 
-export function getWorldId(chainId: ChainId, address: Address): WorldId {
-  return `${chainId}:${getAddress(address)}`;
-}
+export type StoreEventsLog = GetLogsResult<StoreEventsAbi>[number];
+export type BlockLogs = GroupLogsByBlockNumberResult<StoreEventsLog>[number];
 
-export type SchemaToPrimitives<TSchema extends Record<string, SchemaAbiType>> = {
-  [key in keyof TSchema]: SchemaAbiTypeToPrimitiveType<TSchema[key]>;
+export type BaseStorageOperation = {
+  log: StoreEventsLog;
+  namespace: string;
+  name: string;
 };
 
-export function schemaToDefaults<TSchema extends Record<string, SchemaAbiType>>(
-  schema: TSchema
-): SchemaToPrimitives<TSchema> {
-  return Object.fromEntries(
-    Object.entries(schema).map(([key, abiType]) => [key, schemaAbiTypeToDefaultValue[abiType]])
-  ) as SchemaToPrimitives<TSchema>;
-}
+export type SetRecordOperation<TConfig extends StoreConfig> = BaseStorageOperation & {
+  type: "SetRecord";
+} & {
+    [TTable in keyof TConfig["tables"]]: {
+      name: TTable & string;
+      key: Key<TConfig, TTable>;
+      value: Value<TConfig, TTable>;
+    };
+  }[keyof TConfig["tables"]];
+
+export type SetFieldOperation<TConfig extends StoreConfig> = BaseStorageOperation & {
+  type: "SetField";
+} & {
+    [TTable in keyof TConfig["tables"]]: {
+      name: TTable & string;
+      key: Key<TConfig, TTable>;
+    } & {
+      [TValue in keyof Value<TConfig, TTable>]: {
+        fieldName: TValue & string;
+        fieldValue: Value<TConfig, TTable>[TValue];
+      };
+    }[keyof Value<TConfig, TTable>];
+  }[keyof TConfig["tables"]];
+
+export type DeleteRecordOperation<TConfig extends StoreConfig> = BaseStorageOperation & {
+  type: "DeleteRecord";
+} & {
+    [TTable in keyof TConfig["tables"]]: {
+      name: TTable & string;
+      key: Key<TConfig, TTable>;
+    };
+  }[keyof TConfig["tables"]];
+
+export type StorageOperation<TConfig extends StoreConfig> =
+  | SetFieldOperation<TConfig>
+  | SetRecordOperation<TConfig>
+  | DeleteRecordOperation<TConfig>;
